fixup demo site
